Simple Read/Write
Added the ability to send commands to read and write.
added basic text check for setting the global COM default port. It makes sure the final text is in form COM# if a user enters "14" instead of "COM14".

Read:
r32   sends bytes [0x72, 0x32] and the device on the other end would grab those bytes drop the first as a Read flag and then in my case send an I2C message to read register 32.

Write:
w3204 sends bytes [0x77, 0x32, 0x04] and the device on the other end would use 0x77 as a W flag, then I2C write 0x04 to register 32.